requirements.txt
google-genai
feedparser
requests
python-dotenv
